requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
FrEIA @ git+https://github.com/VLL-HD/FrEIA.git@e74b5ec7a10c4f668742de7d639deadfb954ee47
google-auth==1.22.0
google-auth-oauthlib==0.4.1
grpcio==1.32.0
h5py==2.10.0
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
importlib-metadata==2.0.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
multidict==4.7.6
networkx==2.5
numpy==1.19.2
oauthlib==3.1.0
pandas==1.0.3
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymia==0.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.2
SimpleITK==2.0.0
six==1.15.0
tensorboard==2.1.0
tifffile==2020.9.3
torch==1.4.0
torchvision==0.5.0
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.6.0
zipp==3.2.0
